directory sql_models was created, sql_book.py and sql_author.py was moved, main.py and crud.py was refactored
